ER-455 :tractor: - Moving provider interface into providers folder and namespace.
