+++ src/Shared/Recruit.Vacancies.Client/Application/Providers/IGetVacancyTitlesProvider.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Services
+namespace Esfa.Recruit.Vacancies.Client.Application.Providers
 {
     public interface IGetVacancyTitlesProvider
     {
