+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Services/VacancyTitle/VacancyApiTitlesProvider.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Esfa.Recruit.Vacancies.Client.Application.Providers;
 using Esfa.Recruit.Vacancies.Client.Application.Services;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
